+++ MoreShipUpgrades/UpgradeComponents/OneTimeUpgrades/FastEncryption.cs
@@ -13,7 +13,7 @@ namespace MoreShipUpgrades.UpgradeComponents.OneTimeUpgrades
         const float TRANSMIT_MULTIPLIER = 0.2f;
         void Awake()
         {
-            upgradeName = UPGRADE_NAME;
+            upgradeName = UpgradeBus.Instance.PluginConfiguration.OVERRIDE_UPGRADE_NAMES ? UpgradeBus.Instance.PluginConfiguration.FAST_ENCRYPTION_OVERRIDE_NAME : UPGRADE_NAME;
             instance = this;
         }
         public static int GetLimitOfCharactersTransmit(int defaultLimit, string message)