+++ MoreShipUpgrades/UpgradeComponents/OneTimeUpgrades/LockSmith.cs
@@ -24,7 +24,7 @@ namespace MoreShipUpgrades.UpgradeComponents.OneTimeUpgrades
         public int timesStruck;
         void Awake()
         {
-            upgradeName = UPGRADE_NAME;
+            upgradeName = UpgradeBus.Instance.PluginConfiguration.OVERRIDE_UPGRADE_NAMES ? UpgradeBus.Instance.PluginConfiguration.LOCKSMITH_OVERRIDE_NAME : UPGRADE_NAME;
             instance = this;
         }
         internal override void Start()