+++ MoreShipUpgrades/UpgradeComponents/TierUpgrades/BargainConnections.cs
@@ -10,7 +10,7 @@ namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades
         internal const string PRICES_DEFAULT = "225,300,375";
         internal override void Start()
         {
-            upgradeName = UPGRADE_NAME;
+            upgradeName = UpgradeBus.Instance.PluginConfiguration.OVERRIDE_UPGRADE_NAMES ? UpgradeBus.Instance.PluginConfiguration.BARGAIN_CONNECTIONS_OVERRIDE_NAME : UPGRADE_NAME;
             base.Start();
         }
         public static int GetBargainConnectionsAdditionalItems(int defaultAmountItems)