+++ MoreShipUpgrades/UpgradeComponents/TierUpgrades/ChargingBooster.cs
@@ -19,7 +19,7 @@ namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades
         }
         internal override void Start()
         {
-            upgradeName = UPGRADE_NAME;
+            upgradeName = UpgradeBus.Instance.PluginConfiguration.OVERRIDE_UPGRADE_NAMES ? UpgradeBus.Instance.PluginConfiguration.CHARGING_BOOSTER_OVERRIDE_NAME : UPGRADE_NAME;
             chargeCooldown = UpgradeBus.Instance.PluginConfiguration.CHARGING_BOOSTER_COOLDOWN.Value;
             base.Start();
         }