+++ MoreShipUpgrades/UpgradeComponents/TierUpgrades/NightVision.cs
@@ -35,7 +35,7 @@ namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades
         void Awake()
         {
             Instance = this;
-            upgradeName = UPGRADE_NAME;
+            upgradeName = UpgradeBus.Instance.PluginConfiguration.OVERRIDE_UPGRADE_NAMES ? UpgradeBus.Instance.PluginConfiguration.NIGHT_VISION_OVERRIDE_NAME : UPGRADE_NAME;
             nightVisionPrefab = AssetBundleHandler.GetItemObject("Night Vision").spawnPrefab;
         }
         internal override void Start()